hw 8 refactoring
