+++ homework8/src/main/java/by/popkov/homework8/city_database/CityDatabase.java
@@ -1,0 +1,9 @@
+package by.popkov.homework8.city_database;
+
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+
+@Database(entities = {CityEntity.class}, version = 1)
+public abstract class CityDatabase extends RoomDatabase {
+    public abstract CityDao getCityDao();
+}
