+++ homework8/src/main/java/by/popkov/homework8/city_database/CityEntity.java
@@ -1,4 +1,4 @@
-package by.popkov.homework8;
+package by.popkov.homework8.city_database;
 
 import androidx.annotation.NonNull;
 import androidx.room.Entity;
@@ -6,17 +6,17 @@ import androidx.room.Index;
 import androidx.room.PrimaryKey;
 
 @Entity(indices = {@Index(value = {"name"}, unique = true)})
-class CityEntity {
+public class CityEntity {
     @PrimaryKey
     @NonNull
-    String name;
+    public  String name;
 
     @NonNull
-    String getName() {
+    public  String getName() {
         return name;
     }
 
-    CityEntity setName(@NonNull String name) {
+    public  CityEntity setName(@NonNull String name) {
         this.name = name;
         return this;
     }
